backend/requirements.txt
fastapi
uvicorn
networkx
pydantic
python-multipart
requests
google-genai
python-dotenv
